refactor(api-search-index-tasks-ddb-os): inline feature into index.ts
No need for separate feature.ts when index.ts only exports the feature.
