--- packages/api-search-index-tasks-ddb-os/src/feature.ts
@@ -1,15 +1,0 @@
-import { type Container, createFeature } from "@webiny/feature/api";
-import { IndexSettingsManager } from "~/settings/IndexSettingsManager.js";
-import { IndexManagerFactory } from "~/indexManager/IndexManagerFactory.js";
-import { DdbStorageScanner } from "~/storage/StorageScanner.js";
-import { DdbStorageWriter } from "~/storage/StorageWriter.js";
-
-export const SearchIndexTasksDdbOsFeature = createFeature({
-    name: "SearchIndexTasksDdbOs",
-    register(container: Container) {
-        container.register(DdbStorageScanner);
-        container.register(DdbStorageWriter);
-        container.register(IndexSettingsManager);
-        container.register(IndexManagerFactory);
-    }
-});
